+++ validphys2/src/validphys/eff_exponents.py
@@ -11,10 +11,9 @@ import numpy as np
 from reportengine.figure import figuregen
 from reportengine.floatformatting import format_number
 
-from validphys.checks import check_scale, CheckError, make_argcheck, check_positive, check_pdf_normalize_to
-from validphys.plots import BandPDFPlotter
-from validphys.plots import PDFPlotter
-from validphys.pdfbases import (Basis, check_basis)
+from validphys.checks import make_argcheck, check_positive, check_pdf_normalize_to
+from validphys.pdfplots import BandPDFPlotter, PDFPlotter
+from validphys.pdfbases import check_basis
 
 import validphys.pdfgrids as pdfgrids
