--- validphys2/src/validphys/providers.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Wed Mar  9 15:00:35 2016
-
-@author: Zahari Kassabov
-"""
-
-from validphys.results import * #analysis:ignore
-from validphys.plots import *   #analysis:ignore
-from validphys.reweighting import * #analysis:ignore
-
-from validphys.fitdata import replica_data
